pyproject.toml
[tool.poetry]
name = "imgalaxy"
version = "0.1.0"
description = ""
authors = ["Jose Castagnino"]
packages = [{include = "imgalaxy"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
tensorflow = "2.14"
# tensorflow = { version = "^2.14", extras = ["cuda"] }
scikit-learn = "^1.4.0"
tensorflow-datasets = "^4.9.4"
python-decouple = "^3.8"
astropy = "^6.0.0"
importlib-resources = "^6.1.1"
click = "^8.1.7"
matplotlib = "^3.8.2"

[tool.poetry.group.dev.dependencies]
pdbpp = "^0.10.3"
black = "^24.1.1"
pylint = "^3.0.3"
ipython = "^8.21.0"
jupyter = "^1.0.0"
types-requests = "^2.31.0.20240125"
flake8 = "^7.0.0"

[tool.black]
line_length = 88
skip-string-normalization = true

[tool.mypy]
ignore_missing_imports = true
follow_imports = 'skip'

[tool.isort]
multi_line_output = 3
line_length = 88
include_trailing_comma = true

[tool.pylint.master]
jobs = 0
 
[tool.pylint.'MESSAGES CONTROL']
disable = [
  'C',
  'R',
  'fixme',
  'logging-format-interpolation',
  'logging-fstring-interpolation',
  'unspecified-encoding'
]

[tool.pylint.'FORMAT']
max-line-length = 88  # as in black
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
